prototype/requirements.txt
pyTelegramBotAPI
python-i18n